+++ R/AIC.list.R
@@ -63,11 +63,10 @@ AIC.list <- function(object, ..., assess.best=TRUE) {
 #' Extract the best model from a list of models using the AIC.
 #'
 #' @param object the list of models
-#' @param \dots Parameters passed to AIC.list
+#' @param \dots Passed to `AIC()`
 #' @returns The model which is assessed as best.  If more than one are equal,
 #'   the first is chosen.
 #' @export
 get.best.model <- function(object, ...) {
   object[stats::AIC(object, ...)$isBest %in% "Best Model"][[1]]
 }
-