+++ tests/testthat/test-001-add.interval.col.R
@@ -24,7 +24,7 @@ test_that("add.interval.col", {
     regexp="FUN must be a character string or NA",
     info="interval column function must be a character string or NA"
   )
-  
+
   expect_error(
     add.interval.col(name="a", FUN=NA, datatype="interval", desc="test addition"),
     regexp='argument "unit_type" is missing, with no default'
@@ -33,7 +33,7 @@ test_that("add.interval.col", {
     add.interval.col(name="a", FUN=NA, unit_type="foo", datatype="interval", desc="test addition"),
     regexp="should be one of .*inverse_time"
   )
-  
+
   # pretty_name checks
   expect_error(
     add.interval.col(name="a", FUN=NA, unit_type="conc", pretty_name=1:2, datatype="interval", desc=1),
@@ -47,13 +47,13 @@ test_that("add.interval.col", {
     add.interval.col(name="a", FUN=NA, unit_type="conc", pretty_name="", datatype="interval", desc=1),
     regexp="pretty_name must not be an empty string"
   )
-  
+
   expect_error(
     add.interval.col(name="a", FUN=NA, unit_type="conc", pretty_name="a", datatype="individual"),
     regexp="Only the 'interval' datatype is currently supported.",
     info="interval column datatype must be 'interval'"
   )
-  
+
   expect_error(
     add.interval.col(name="a", FUN=NA, unit_type="conc", pretty_name="a", datatype="interval", desc=1:2),
     regexp="desc must have length == 1",
@@ -69,7 +69,7 @@ test_that("add.interval.col", {
     regexp="'depends' must be NULL or a character vector",
     info="depends column must be a NULL or a character string"
   )
-  
+
   expect_error(
     add.interval.col(name="a", FUN="this function does not exist", unit_type="conc", pretty_name="foo", datatype="interval", desc="test addition"),
     regexp="The function named '.*' is not defined.  Please define the function before calling add.interval.col.",
@@ -99,7 +99,7 @@ test_that("add.interval.col", {
     regexp="All names for the formalsmap list must be arguments to the function",
     info="formalsmap arguments must map to function arguments"
   )
-  
+
   expect_equal(
     {
       add.interval.col(name="a", FUN=NA, unit_type="conc", pretty_name="a", datatype="interval", desc="test addition")
@@ -114,7 +114,9 @@ test_that("add.interval.col", {
       sparse=FALSE,
       formalsmap=list(),
       depends=NULL,
-      datatype="interval"
+      datatype="interval",
+      pptestcd_cdisc="a",
+      pptest_cdisc="test addition"
     ),
     info="interval column assignment works with FUN=NA"
   )
@@ -132,7 +134,9 @@ test_that("add.interval.col", {
       sparse=FALSE,
       formalsmap=list(),
       depends=NULL,
-      datatype="interval"
+      datatype="interval",
+      pptestcd_cdisc="a",
+      pptest_cdisc="test addition"
     ),
     info="interval column assignment works with FUN=a character string"
   )
@@ -150,7 +154,9 @@ test_that("add.interval.col", {
       sparse=FALSE,
       formalsmap=list(x="values"),
       depends=NULL,
-      datatype="interval"
+      datatype="interval",
+      pptestcd_cdisc="a",
+      pptest_cdisc="test addition"
     ),
     info="interval column assignment works with FUN=NA"
   )
@@ -170,10 +176,37 @@ test_that("fake parameters", {
     depends="does_not_exist"
   )
   expect_error(
-    sort.interval.cols(),
+    sort_interval_cols(),
     regexp="Invalid dependencies for interval column (please report this as a bug): fake_parameter The following dependencies are missing: does_not_exist",
     fixed=TRUE
   )
+})
+
+test_that("add.interval.col rejects invalid pptestcd_cdisc types", {
+  expect_error(
+    add.interval.col(name="a", FUN="mean", unit_type="conc", pretty_name="a",
+                     desc="test", pptestcd_cdisc=123),
+    regexp="pptestcd_cdisc must be a character string or a list"
+  )
+})
+
+test_that("add.interval.col rejects invalid pptest_cdisc types", {
+  expect_error(
+    add.interval.col(name="a", FUN="mean", unit_type="conc", pretty_name="a",
+                     desc="test", pptest_cdisc=123),
+    regexp="pptest_cdisc must be a character string or a list"
+  )
+})
+
+test_that("add.interval.col accepts list for pptestcd_cdisc", {
+  add.interval.col(name="a", FUN="mean", unit_type="conc", pretty_name="a",
+                   desc="test",
+                   pptestcd_cdisc=list(route=list(extravascular="EV", intravascular="IV")),
+                   pptest_cdisc="test desc")
+  result <- get("interval.cols", envir=PKNCA:::.PKNCAEnv)[["a"]]
+  expect_true(is.list(result$pptestcd_cdisc))
+  expect_equal(result$pptestcd_cdisc$route$extravascular, "EV")
+  expect_equal(result$pptestcd_cdisc$route$intravascular, "IV")
 })
 
 # Reset the original state